refactor yvoty folder tests
