+++ tests/integration/desafios/segundoCiclo/Yvoty/SuperYvoty1-test.js
@@ -1,4 +1,5 @@
-import { moduloActividad } from '../../../helpers/actividadTest';
+import { moduloActividad } from '../../../../helpers/actividadTest';
+
 
 const nombre = 'SuperYvoty1';
