+++ tests/integration/desafios/segundoCiclo/Yvoty/SuperYvoty2-test.js
@@ -1,4 +1,4 @@
-import { moduloActividad } from '../../../helpers/actividadTest';
+import { moduloActividad } from '../../../../helpers/actividadTest';
 
 const nombre = 'SuperYvoty2';
