+++ tests/integration/desafios/segundoCiclo/Yvoty/YvotyDespiertaMasLuciernagas-test.js
@@ -1,4 +1,4 @@
-import { moduloActividad } from '../../../helpers/actividadTest';
+import { moduloActividad } from '../../../../helpers/actividadTest';
 
 const nombre = 'YvotyDespiertaMasLuciernagas';
